+++ backend/users/Touir.ExpensesManager.Users/Infrastructure/EmailHTMLTemplate.cs
@@ -10,5 +10,14 @@
                 public static readonly string VerificationLink = "VERIFICATION_LINK";
             }
         }
+
+        public static class PasswordReset
+        {
+            public static readonly string Key = "PASSWORD_RESET_TEMPLATE";
+            public static class Variables
+            {
+                public static readonly string ResetLink = "RESET_LINK";
+            }
+        }
     }
 }